Auto setup of project

+++ make_project.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 from __future__ import print_function
 import shutil
-
+import subprocess
 import begin
 import jinja2
 import os
@@ -40,7 +40,7 @@ def copyfile(src, dst, follow_symlinks=False):
         shutil.copy(src,dst)
 
 @begin.start
-def main(name, language, root=None):
+def main(name, language, project_root=None):
     language_dir = os.path.join(languages_dir, language)
 
     if not os.path.exists(language_dir):
@@ -49,12 +49,12 @@ def main(name, language, root=None):
               file=sys.stderr)
         sys.exit(1)
 
-    if root is None:
+    if project_root is None:
         mkdir(name, exist_ok=True)
-        root = name
+        project_root = name
 
     language_dir = os.path.realpath(language_dir)
-    root = os.path.realpath(root)
+    project_root = os.path.realpath(project_root)
 
     project = Project(name, language)
 
@@ -66,40 +66,51 @@ def main(name, language, root=None):
         , undefined=StrictUndefined
     )
 
-    for base, dirs, files in os.walk(language_dir):
+    template_files_dir = os.path.join(language_dir, 'files')
 
-        rdir = os.path.relpath(base, language_dir)
-        ndir = os.path.join(root, rdir)
-        os.chdir(ndir)
+    if os.path.exists(template_files_dir):
+        for base, dirs, files in os.walk(template_files_dir):
 
-        for d in dirs:
-            mkdir(d, exist_ok=True)
+            rdir = os.path.relpath(base, template_files_dir)
+            ndir = os.path.join(project_root, rdir)
+            os.chdir(ndir)
 
-        for file in files:
-            src = os.path.join(language_dir, rdir, file)
-            se = os.path.splitext(file)
-            if se[1] == '.jinja':
-                dest = os.path.join(os.getcwd(), se[0])
-                if os.path.exists(dest):
-                    os.unlink(dest)
-                print("Expanding {} -> {}".format(src, dest))
-                template = env.get_template(src)
+            for d in dirs:
+                mkdir(d, exist_ok=True)
 
-                try:
-                    out = template.render(project=project)
-                    out_path = path.join(dest)
-                    with open(out_path, "w") as out_fh:
-                        print(out, file=out_fh)
+            for file in files:
+                src = os.path.join(template_files_dir, rdir, file)
+                se = os.path.splitext(file)
+                if se[1] == '.jinja':
+                    dest = os.path.join(os.getcwd(), se[0])
+                    if os.path.exists(dest):
+                        os.unlink(dest)
+                    print("Expanding {} -> {}".format(src, dest))
+                    template = env.get_template(src)
 
-                except jinja2.exceptions.UndefinedError as ex:
-                    print("\nERROR: Undefined attribute: {}".format(ex.message), file=sys.stderr)
-                    print(
-                        "You can define template variables using a semi-colon-separated list of <KEY>=<VALUE> as the value of the --defines flag.",
-                        file=sys.stderr)
+                    try:
+                        out = template.render(project=project)
+                        out_path = path.join(dest)
+                        with open(out_path, "w") as out_fh:
+                            print(out, file=out_fh)
 
-            else:
-                dest = os.path.join(os.getcwd(), file)
-                if os.path.lexists(dest):
-                    os.unlink(dest)
-                print("Copying {} -> {}".format(src, dest))
-                copyfile(src, dest, follow_symlinks=False)
+                    except jinja2.exceptions.UndefinedError as ex:
+                        print("\nERROR: Undefined attribute: {}".format(ex.message), file=sys.stderr)
+                        print(
+                            "You can define template variables using a semi-colon-separated list of <KEY>=<VALUE> as the value of the --defines flag.",
+                            file=sys.stderr)
+
+                else:
+                    dest = os.path.join(os.getcwd(), file)
+                    if os.path.lexists(dest):
+                        os.unlink(dest)
+                    print("Copying {} -> {}".format(src, dest))
+                    copyfile(src, dest, follow_symlinks=False)
+
+        setup_file = path.join(language_dir, 'setup.py') if path.exists(path.join(language_dir, 'setup.py')) else \
+            path.join(language_dir, 'setup.sh') if path.exists(path.join(language_dir, 'setup.sh')) else \
+            path.join(language_dir, 'setup') if path.exists(path.join(language_dir, 'setup')) else None
+
+        if setup_file:
+            print("Running {}".format(setup_file))
+            subprocess.call(setup_file, env={'PATH':os.getenv('PATH'), 'PROJECT_DIR': project_root}, cwd=project_root)
